backend/requirements.txt
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.0
pydantic-settings>=2.0
openai>=1.0
anthropic>=0.20
asyncpg>=0.29
psycopg2-binary>=2.9
pgvector>=0.2
sqlalchemy>=2.0
python-dotenv>=1.0
httpx>=0.27
python-jose[cryptography]>=3.3
requests>=2.31
beautifulsoup4>=4.12
celery>=5.3
redis>=5.0
pdfminer.six>=20221105
